# Reuse the BlackFriday instance when possible