--- a/hugolib/config.go
+++ b/hugolib/config.go
@@ -210,7 +210,7 @@ func loadDefaultSettingsFor(v *viper.Viper) error {
 	v.SetDefault("paginate", 10)
 	v.SetDefault("paginatePath", "page")
 	v.SetDefault("summaryLength", 70)
-	v.SetDefault("blackfriday", c.NewBlackfriday())
+	v.SetDefault("blackfriday", c.BlackFriday)
 	v.SetDefault("rSSUri", "index.xml")
 	v.SetDefault("rssLimit", -1)
 	v.SetDefault("sectionPagesMenu", "")